+++ src/Modulator/GSM/BCJR/Modulator_GSM_BCJR.hxx
@@ -130,7 +130,7 @@ void Modulator_GSM_BCJR<Q,MAX>
 {
 	std::fill(this->symb_apriori_prob.begin(), this->symb_apriori_prob.end(), 0);
 
-	for (auto i = 0; i < this->symb_apriori_prob.size() / m_order; i++)
+	for (auto i = 0; i < (int) this->symb_apriori_prob.size() / m_order; i++)
 		for (auto j = 0; j < this->m_order; j++)
 			for (auto k = 0; k < this->n_bits_per_symb; k++)
 			{
